requirements.txt
ttkbootstrap
Pillow
face-recognition
opencv-python
ultralytics
matplotlib
